requirements.txt
# Django Framework (Python 3.14 supports Django 5.x)
django>=5.1,<6.0

# Authentication & Authorization
django-allauth>=0.63.0

# Database
psycopg[binary]>=3.1.0  # PostgreSQL adapter for Django 5.x

# Environment Management
python-dotenv>=1.0.0

# Image Processing
pillow>=10.0.0

# Additional Utilities
requests>=2.31.0  # HTTP library (required by allauth)
django-ratelimit>=4.1.0  # Rate limiting for security
django-axes>=6.1.0  # Login attempt tracking
redis>=4.5.0  # Persistent cache for rate limiting
